+++ Hollow Knight/Assets/Animation/Monster/Husk Hornhead/Hit.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: Die
+  m_Name: Hit
   serializedVersion: 6
   m_Legacy: 0
   m_Compressed: 0
@@ -20,21 +20,9 @@ AnimationClip:
   m_PPtrCurves:
   - curve:
     - time: 0
-      value: {fileID: -9140909, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
+      value: {fileID: 634174985, guid: cf3c005e2b667e346a9e70ee0a103ca0, type: 3}
     - time: 0.083333336
-      value: {fileID: 1565816745, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - time: 0.16666667
-      value: {fileID: 1367032097, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - time: 0.25
-      value: {fileID: -1934673095, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - time: 0.33333334
-      value: {fileID: -312847888, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - time: 0.41666666
-      value: {fileID: -530777868, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - time: 0.5
-      value: {fileID: 690801324, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - time: 0.5833333
-      value: {fileID: -1117101311, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
+      value: {fileID: 1726007114, guid: cf3c005e2b667e346a9e70ee0a103ca0, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
@@ -54,25 +42,19 @@ AnimationClip:
       customType: 23
       isPPtrCurve: 1
     pptrCurveMapping:
-    - {fileID: -9140909, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - {fileID: 1565816745, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - {fileID: 1367032097, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - {fileID: -1934673095, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - {fileID: -312847888, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - {fileID: -530777868, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - {fileID: 690801324, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
-    - {fileID: -1117101311, guid: c86e0f5e9107a874fb64efaad8f9a04d, type: 3}
+    - {fileID: 634174985, guid: cf3c005e2b667e346a9e70ee0a103ca0, type: 3}
+    - {fileID: 1726007114, guid: cf3c005e2b667e346a9e70ee0a103ca0, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.6666666
+    m_StopTime: 0.16666667
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
     m_HasAdditiveReferencePose: 0
-    m_LoopTime: 0
+    m_LoopTime: 1
     m_LoopBlend: 0
     m_LoopBlendOrientation: 0
     m_LoopBlendPositionY: 0
