+++ Hollow Knight/Assets/Script/Monster/Husk Hornhead/HuskHornheadAttack.cs
@@ -2,14 +2,13 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class WanderingHuskWalk : MonsterWalk
+public class HuskHornheadAttack : MonsterAttack
 {
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         base.OnStateEnter(animator, stateInfo, layerIndex);
-        monster = animator.GetComponent<WanderingHusk>();
+        monster = animator.GetComponent<HuskHornhead>();
     }
-
     override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         base.OnStateUpdate(animator, stateInfo, layerIndex);
