+++ Hollow Knight/Assets/Script/Monster/Husk Hornhead/HuskHornheadIdle.cs
@@ -1,17 +1,19 @@
 using System.Collections;
 using System.Collections.Generic;
+using System.Threading;
 using UnityEngine;
 
-public class WanderingHuskWalk : MonsterWalk
+public class HuskHornheadIdle : MonsterIdle
 {
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         base.OnStateEnter(animator, stateInfo, layerIndex);
-        monster = animator.GetComponent<WanderingHusk>();
+        monster = animator.GetComponent<HuskHornhead>();
     }
-
     override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         base.OnStateUpdate(animator, stateInfo, layerIndex);
     }
+
+   
 }
