+++ Hollow Knight/Assets/Script/Monster/Wandering Husk/WanderingHuskIdle.cs
@@ -5,7 +5,6 @@ using UnityEngine;
 
 public class WanderingHuskIdle : MonsterIdle
 {
-    private WanderingHusk monster;
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         base.OnStateEnter(animator, stateInfo, layerIndex);
@@ -14,21 +13,8 @@ public class WanderingHuskIdle : MonsterIdle
 
     override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        Attack(animator);
         base.OnStateUpdate(animator, stateInfo, layerIndex);
     }
 
-    private void Attack(Animator anim)
-    {
-        Vector2 direction = monster.target.position - monsterTransform.position;
-        float distance = direction.magnitude;
-        if (distance <= monster.attackRange)
-        {
-            anim.SetBool("Idle", false);
-            anim.SetBool("Walk", false);
-            anim.SetTrigger("Hit");
-            monster.Flip(monster.target.position.x, monsterTransform.position.x);
-        }
-        return;
-    }
+    
 }
